Added Pin + Positioning
Introduced a pin (#pin) as an image nested in a div (#mappin) alongside the map (#campus). The Pin is positioned absolutely with respect to the "mappin" div. It's location can be adjusted by altering the top/right/bottom/left CSS properties of #pin.

Hopefully this method can be used to drop a pin on the user's click via JavaScript

-Pin's point isn't necessarily correspondent to the location of the image, since it's more to the bottom left.