Score Calculation v2
Adjusted Score calculation to make all values from 0-100 possible.

Before: Distance of error was scaled to the maximum possible line on the map (diagonal), but a user could never practically be that far, resulting in a minimum score of 14-30.

After: Distance of error is scaled to the maximum possible error for the given location (location->farthest corner), so a user can actually guess on the lowest end of the scale, resulting in a minimum score of 0-1.